@@ -9,19 +9,29 @@ import { AlbumService } from 'src/app/services/album.service';
 import { SongService } from 'src/app/services/song.service';
 import { ISong } from 'src/app/interfaces/ISong';
 import { SongLatestComponent } from "./song/song-latest/song-latest.component";
+import { MatButtonModule } from '@angular/material/button';
+import { MatDialog, MatDialogRef, MatDialogModule } from '@angular/material/dialog';
+import { Router } from '@angular/router';
 
 @Component({
-    selector: 'app-home-page',
-    standalone: true,
-    template: `
+  selector: 'app-home-page',
+  standalone: true,
+  template: `
     <div class="library-component">
-      <h1 class="library-component-title">Add album</h1>
-      <div #listAlbums class="library-component-album-list">
+      <div class="button-text">
+        <h1 class="library-component-title">Add album</h1>
+        <i id="album-button" class="material-icons w3-xxlarge" (click)="addSong()">add</i>
+      </div>
+        <div #listAlbums class="library-component-album-list">
         <app-album class="album-component" *ngFor="let album of this.albumList" [album]="album"></app-album>
       </div>
     </div>
     <div class = "library-component">
+    <div class="button-text">
       <h1 class = "library-component-title">Add song</h1>
+      <i id="song-button" class="material-icons w3-xxlarge" (click)="addAlbum()">add</i>
+      </div>
+
       <div class = "library-component-song-list">
         <app-song class = "song-component" *ngFor="let song of this.songList" [song]="song"></app-song>
       </div>
@@ -39,54 +49,63 @@ import { SongLatestComponent } from "./song/song-latest/song-latest.component";
       </div>
     </div>
   `,
-    styleUrls: ['./home-page.component.css'],
-    imports: [CommonModule, AlbumComponent, SongComponent, SongLatestComponent]
+  styleUrls: ['./home-page.component.css'],
+  imports: [CommonModule, AlbumComponent, SongComponent, SongLatestComponent, MatDialogModule]
+
 })
 export class HomePageComponent implements AfterViewChecked {
-  @ViewChild('listAlbums', {static:false}) listAlbums! : ElementRef;
-  @ViewChild('listSongLatest', {static:false}) listSongLatest! : ElementRef;
+  @ViewChild('listAlbums', { static: false }) listAlbums!: ElementRef;
+  @ViewChild('listSongLatest', { static: false }) listSongLatest!: ElementRef;
   @ViewChild('listSongOutStanding', {static:false}) listSongOutStanding! : ElementRef;
 
   albumService: AlbumService = inject(AlbumService);
   songService: SongService = inject(SongService);
   cookieService: CookieService = inject(CookieService);
 
-  albumList : IAlbum[] = [];
-  songList : ISong[] = [];
+  albumList: IAlbum[] = [];
+  songList: ISong[] = [];
   songListLastes : ISong[] = [];
+
     
-  constructor(){
+  constructor(private router: Router,public dialog: MatDialog) {
     this.getAlbums();
     this.getSongs();
     this.getSongsByDate();
   }
 
-  async getAlbums(){
+  async getAlbums() {
     await this.albumService.getAlbums(this.cookieService.get("USERID"))
-    .then((value:IResponse<IAlbum>) => {
-      value.data.forEach(element => {
-        this.albumList.push(element);
-      });
-    })
-    .catch(error => {
-      console.error(error);
-    })
-    
+      .then((value: IResponse<IAlbum>) => {
+        value.data.forEach(element => {
+          this.albumList.push(element);
+        });
+
+        console.log(value.Result)
+        if (value.Result.statuscode === "403") {
+          this.router.navigate(['/login']);
+        }
+       
+
+      })
+      .catch(error => {
+        console.error(error);
+      })
+
   }
 
-  async getSongs(){
+  async getSongs() {
     await this.songService.getSongs()
-    .then((value:IResponse<ISong>) =>{
-      value.data.forEach(element => {
-        this.songList.push(element);
+      .then((value: IResponse<ISong>) => {
+        value.data.forEach(element => {
+          this.songList.push(element);
+        });
+      })
+      .catch(err => {
+        console.error(err);
       });
-    })
-    .catch(err => {
-      console.error(err);
-    });
   }
 
-  async getSongsByDate(){
+  async getSongsByDate() {
     await this.songService.getSongsByDate()
     .then((value:IResponse<ISong>) =>{
       value.data.forEach(element => {
@@ -99,13 +118,13 @@ export class HomePageComponent implements AfterViewChecked {
       console.error(err);
     });
   }
-  
+
 
   ngAfterViewChecked() {
     const albumList = this.listAlbums.nativeElement;
     const albums = albumList.querySelectorAll('.album-component');
     const entireAlbums = albums.length;
-    
+
     const songLatestList = this.listSongLatest.nativeElement;
     const songsLatests = songLatestList.querySelectorAll('.songLatest-component');
     const entireSongLatests = songsLatests.length;
@@ -121,6 +140,8 @@ export class HomePageComponent implements AfterViewChecked {
     this.countColumns(containerAlbum,entireAlbums,50,2);
     this.countColumns(containerSongLatest, entireSongLatests,100,4);
     this.countColumns(containerSongOutStanding,entireSongOutStanding,100,4);
+    this.countColumns(containerAlbum, entireAlbums, 50, 2);
+    this.countColumns(containerSongLatest, entireSongLatests, 100, 4);
   }
 
   countColumns(container : any, entire : number, capacity:number, columns:number ){
@@ -129,4 +150,87 @@ export class HomePageComponent implements AfterViewChecked {
       container.style.gridTemplateColumns = `repeat(${numColumns}, ${capacity}%)`;
     }
   }
+
+
+  addAlbum() {
+    this.openSongDialog('0ms', '0ms')
+  }
+  addSong() {
+    this.openAlbumDialog('0ms', '0ms')
+
+  }
+
+  openSongDialog(enterAnimationDuration: string, exitAnimationDuration: string): void {
+    this.dialog.open(CreateSongDialog, {
+      width: '250px',
+      enterAnimationDuration,
+      exitAnimationDuration,
+    });
+  }
+
+  openAlbumDialog(enterAnimationDuration: string, exitAnimationDuration: string): void {
+    this.dialog.open(CreateAlbumDialog, {
+      width: '250px',
+      enterAnimationDuration,
+      exitAnimationDuration,
+    });
+  }
 }
+
+@Component({
+  selector: 'CreateSongDialog',
+  template: `
+  <h1 mat-dialog-title>CreateSongDialog</h1>
+  <div mat-dialog-content>
+    Would you like to delete cat.jpeg?
+  </div>
+  <div mat-dialog-actions>
+    <button mat-button mat-dialog-close>No</button>
+    <button mat-button mat-dialog-close cdkFocusInitial>Ok</button>
+</div>
+`,
+  standalone: true,
+  imports: [MatDialogModule, MatButtonModule],
+})
+export class CreateSongDialog {
+  constructor(public dialogRef: MatDialogRef<CreateSongDialog>) { }
+}
+
+
+@Component({
+  selector: 'CreateAlbumDialog',
+  template: `
+  <h1 mat-dialog-title>CreateAlbumDialog</h1>
+  <div mat-dialog-content>
+  <input type="text" id="albumTitle" placeholder="Album Title" >
+  <input type="text" id="albumDescription" placeholder="Album Description" >
+
+  </div>
+  <div mat-dialog-actions>
+    <button mat-button mat-dialog-close>Cancel</button>
+    <button mat-button mat-dialog-close cdkFocusInitial (click)="createAlbum()">Create</button>
+</div>
+`,
+  standalone: true,
+  imports: [MatDialogModule, MatButtonModule],
+})
+export class CreateAlbumDialog {
+  constructor(public dialogRef: MatDialogRef<CreateAlbumDialog>) { }
+  albumService: AlbumService = inject(AlbumService);
+  songService: SongService = inject(SongService);
+  cookieService: CookieService = inject(CookieService);
+
+  createAlbum() {
+
+    const data: IAlbum = {
+      idAlbum: 1,
+      titulo: "",
+      descripcion: '',
+      fechaCreacion: "1991-07-21",
+      idUsuario: 26
+    }
+    this.albumService.createAlbum(data.idUsuario, data);
+  }
+}
+
+
