@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { Component, inject } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { LoginComponent } from './login/login.component';
 import { RouterOutlet } from '@angular/router';
@@ -6,13 +6,14 @@ import { Router } from '@angular/router';
 
 import { AlertComponent } from '../alert/alert.component';
 import { AlertInterface } from '../interfaces/IAlert';
+import { CookieService } from '../services/cookie.service';
 
 @Component({
   selector: 'app-home',
   imports: [CommonModule, RouterOutlet, LoginComponent, AlertComponent],
   standalone: true,
   template: `
-      <div id="home-component-header">
+      <div id="home-component-header" *ngIf="this.loginFlag">
         <i id="home-button" class="material-icons w3-xxlarge" (click)="homeButton()">home</i>
         <i id="user-button"class="material-icons w3-xxlarge" (click)="userButton()">person</i>
       </div>
@@ -23,11 +24,20 @@ import { AlertInterface } from '../interfaces/IAlert';
 })
 export class HomeComponent {
   alerts: AlertInterface[];
-
   alert = {} as AlertInterface;
+
+  cookie!: string;
+  cookieService: CookieService = inject(CookieService);
+  loginFlag: boolean = false
 
   constructor(private router: Router) {
     this.alerts = [] as AlertInterface[];
+    this.cookie = this.cookieService.get("SESSIONID");
+    this.loginFlag = false;
+
+  }
+
+  ngOnInit() {
   }
 
   addAlert(alert: AlertInterface) {
@@ -36,6 +46,8 @@ export class HomeComponent {
 
     console.log(this.alert);
   }
+
+
 
   myFunction() {
     this.alerts.pop();
