@@ -2,6 +2,6 @@ export interface IAlbum {
     idAlbum:number,
     titulo:string,
     descripcion:string,
-    fechaCreacion:Date,
+    fechaCreacion:string,
     idUsuario:number,
 }
